Refactor solution
This commit firstly renames the "DocumentFormat" namespace component into
"ContractArchitect", so DocumentFormat.OpenXml.Extensions now becomes
ContractArchitect.OpenXml.Extensions, for example. This was done because
Microsoft took the decision to keep the Open XML SDK very low level. Therefore,
it did not make sense to use the "DocumentFormat" namespace for these higher
level extensions (in particular also the transformations).

Secondly, this commit removes the OpenXmlExtensionsTest project, because that
actually only tests the enhancements contributed to the Open-XML-SDK project
itself. They will be moved to their own project on GitHub.

Thirdly, the exploratory PackagingExtensions project was also removed, because
it is currently not used. It will also be moved to its own project on GitHub.