Fix Binary to Open XML Translator
The Binary to Open XML Translator produced partly wrong Open XML markup. This
commit fixes that. In addition, this commit fixes related issues in Contract
Architect.

The Web Application used a fixed template for transforming house style documents
to HTML. This commit changes that behavior so that the uploaded document is
used as a template and style separators are inserted if the uploaded document
contains style separators as well.