Use new Open XML SDK with System.IO.Packaging
This commit makes the changes necessary for using the new Open XML SDK with
System.IO.Packaging (the latter replacing WindowsBase).

The switch to the new SDK and in particular the new System.IO.Packaging
introduced a bug which meant that the HosueStyleTransformation could no longer
transform certain documents. This was due to a call to document.Package.Flush()
in an extension method within WordprocessingDocumentExtensions.cs. This commit
removes that call (which was also probably unnecessary).